Added Entity, Repository, Service, and Controller for 'Account' operations
