+++ src/main/java/com/amigoscode/group/ebankingsuite/Account/AccountRepository.java
@@ -1,0 +1,11 @@
+package com.amigoscode.group.ebankingsuite.Account;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface AccountRepository extends JpaRepository<Account, Integer> {
+    List<Account> findByUserId(Integer userId);
+}
