+++ src/main/java/com/amigoscode/group/ebankingsuite/EBankingSuiteApplication.java
@@ -1,0 +1,12 @@
+package com.amigoscode.group.ebankingsuite;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EBankingSuiteApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(EBankingSuiteApplication.class, args);
+    }
+
+}
