+++ src/main/java/com/amigoscode/group/ebankingsuite/exception/InvalidAuthenticationException.java
@@ -1,0 +1,8 @@
+package com.amigoscode.group.ebankingsuite.exception;
+
+public class InvalidAuthenticationException extends RuntimeException{
+
+    public InvalidAuthenticationException(String message) {
+        super(message);
+    }
+}
