+++ src/main/java/com/amigoscode/group/ebankingsuite/exception/InvalidAuthorizationException.java
@@ -1,0 +1,8 @@
+package com.amigoscode.group.ebankingsuite.exception;
+
+public class InvalidAuthorizationException extends RuntimeException{
+
+    public InvalidAuthorizationException(String message) {
+        super(message);
+    }
+}
