+++ src/main/java/com/amigoscode/group/ebankingsuite/exception/ResourceExistsException.java
@@ -1,0 +1,8 @@
+package com.amigoscode.group.ebankingsuite.exception;
+
+public class ResourceExistsException extends RuntimeException{
+
+    public ResourceExistsException(String message) {
+        super(message);
+    }
+}
