+++ src/main/java/com/amigoscode/group/ebankingsuite/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.amigoscode.group.ebankingsuite.exception;
+
+public class ResourceNotFoundException extends RuntimeException{
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
