+++ src/main/java/com/amigoscode/group/ebankingsuite/universal/ApiResponse.java
@@ -1,0 +1,22 @@
+package com.amigoscode.group.ebankingsuite.universal;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class ApiResponse {
+
+    private String message;
+    private Object data;
+
+    public ApiResponse(String message, String data) {
+        this.message = message;
+        this.data = data;
+    }
+
+    public ApiResponse(String message) {
+        this.message = message;
+    }
+}
