+++ src/main/java/com/amigoscode/group/ebankingsuite/user/UserRepository.java
@@ -1,0 +1,12 @@
+package com.amigoscode.group.ebankingsuite.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,String> {
+
+    boolean existsByEmailAddress(String emailAddress);
+    Optional<User> findByEmailAddress(String email);
+}
