+++ src/main/java/com/amigoscode/group/ebankingsuite/user/resquests/UserAuthenticationRequests.java
@@ -1,0 +1,7 @@
+package com.amigoscode.group.ebankingsuite.user.resquests;
+
+public record UserAuthenticationRequests(
+        String emailAddress,
+        String password
+) {
+}
