+++ src/main/java/com/amigoscode/group/ebankingsuite/user/resquests/UserRegistrationRequest.java
@@ -1,0 +1,8 @@
+package com.amigoscode.group.ebankingsuite.user.resquests;
+
+public record UserRegistrationRequest(
+        String fullName,
+        String emailAddress,
+        String password
+) {
+}
