--- src/main/java/com/amogoscode/group/ebankingsuite/EBankingSuiteApplication.java
@@ -1,12 +1,0 @@
-package com.amogoscode.group.ebankingsuite;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EBankingSuiteApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(EBankingSuiteApplication.class, args);
-    }
-
-}
