--- src/main/java/com/amogoscode/group/ebankingsuite/exception/InvalidAuthenticationException.java
@@ -1,8 +1,0 @@
-package com.amogoscode.group.ebankingsuite.exception;
-
-public class InvalidAuthenticationException extends RuntimeException{
-
-    public InvalidAuthenticationException(String message) {
-        super(message);
-    }
-}
