--- src/main/java/com/amogoscode/group/ebankingsuite/exception/InvalidAuthorizationException.java
@@ -1,8 +1,0 @@
-package com.amogoscode.group.ebankingsuite.exception;
-
-public class InvalidAuthorizationException extends RuntimeException{
-
-    public InvalidAuthorizationException(String message) {
-        super(message);
-    }
-}
