--- src/main/java/com/amogoscode/group/ebankingsuite/exception/ResourceExistsException.java
@@ -1,8 +1,0 @@
-package com.amogoscode.group.ebankingsuite.exception;
-
-public class ResourceExistsException extends RuntimeException{
-
-    public ResourceExistsException(String message) {
-        super(message);
-    }
-}
