--- src/main/java/com/amogoscode/group/ebankingsuite/exception/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package com.amogoscode.group.ebankingsuite.exception;
-
-public class ResourceNotFoundException extends RuntimeException{
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
