--- src/main/java/com/amogoscode/group/ebankingsuite/universal/ApiResponse.java
@@ -1,22 +1,0 @@
-package com.amogoscode.group.ebankingsuite.universal;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class ApiResponse {
-
-    private String message;
-    private Object data;
-
-    public ApiResponse(String message, String data) {
-        this.message = message;
-        this.data = data;
-    }
-
-    public ApiResponse(String message) {
-        this.message = message;
-    }
-}
