--- src/main/java/com/amogoscode/group/ebankingsuite/user/resquests/UserAuthenticationRequests.java
@@ -1,7 +1,0 @@
-package com.amogoscode.group.ebankingsuite.user.resquests;
-
-public record UserAuthenticationRequests(
-        String emailAddress,
-        String password
-) {
-}
