--- src/main/java/com/amogoscode/group/ebankingsuite/user/resquests/UserRegistrationRequest.java
@@ -1,8 +1,0 @@
-package com.amogoscode.group.ebankingsuite.user.resquests;
-
-public record UserRegistrationRequest(
-        String fullName,
-        String emailAddress,
-        String password
-) {
-}
