+++ src/test/java/com/amigoscode/group/ebankingsuite/EBankingSuiteApplicationTests.java
@@ -1,0 +1,13 @@
+package com.amigoscode.group.ebankingsuite;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EBankingSuiteApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
