--- src/test/java/com/amogoscode/group/ebankingsuite/EBankingSuiteApplicationTests.java
@@ -1,13 +1,0 @@
-package com.amogoscode.group.ebankingsuite;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EBankingSuiteApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
